Vehiculo

+++ st-parranderos-jdo-est/st-parranderos-jdo-est/src/main/java/uniandes/isis2304/parranderos/interfazApp/DialogoAdicionarAscensor.java
@@ -203,7 +203,7 @@ public class DialogoAdicionarAscensor extends JDialog implements ActionListener
 				{
 					int capacidadNormal = Integer.parseInt(capacidadNormalStr);
 					double pesoMaximo = Double.parseDouble(pesoMaximoStr);
-					double area = Double.parseDouble(areaStr);
+					long area = Long.parseLong(areaStr);
 					if ( interfaz.buscarAreaPorValor(area) == null)
 					{
 						interfaz.adicionarArea();
